feat(datasets): show example details in a slide-over
